+++ index.js
@@ -2,7 +2,7 @@ const express = require("express");
 const app = express();
 const cors = require("cors");
 const PORT = process.env.PORT || 3500;
-
+const corsOptions = require("./config/corsOptions");
 const { logger } = require("./middleware/logEvents");
 const errorHandler = require("./middleware/errorHandler");
 
@@ -11,22 +11,6 @@ const path = require("path");
 //custom middleware for to listen to events
 app.use(logger);
 
-const whitelist = [
-  "https://www.google.com",
-  "http://127:0:0:1:5500",
-  "http://localhost:3500",
-];
-
-const corsOptions = {
-  origin: (origin, callback) => {
-    if (whitelist.indexOf(origin) !== -1 || !origin) {
-      callback(null, true);
-    } else {
-      callback(new Error("Not allowed by CORS"));
-    }
-  },
-  optionsSuccessStatus: 200,
-};
 app.use(cors(corsOptions));
 
 //built-in middleware to handle urlencoded data
