+++ routes/api/employees.js
@@ -1,26 +1,15 @@
 const express = require("express");
 const router = express.Router();
 const path = require("path");
-const data = {};
-data.employees = require("../../data/employees.json");
+const employeesController = require("../../controllers/employeesController");
 
 router
   .route("/")
-  .get((req, res) => {
-    res.json(data.employees);
-  })
-  .post((req, res) => {
-    req.json({ firstname: req.body.firstname, lastname: req.body.lastname });
-  })
-  .put((req, res) => {
-    req.json({ firstname: req.body.firstname, lastname: req.body.lastname });
-  })
-  .delete((req, res) => {
-    req.json({ firstname: req.body.id });
-  });
+  .get(employeesController.getAllEmployees)
+  .post(employeesController.createNewEmployee)
+  .put(employeesController.updateNewEmployee)
+  .delete(employeesController.deleteEmployee);
 
-router.route("/:id").get((req, res) => {
-  res.json({ id: req.params.id });
-});
+router.route("/:id").get(employeesController.getEmployee);
 
 module.exports = router;
